tests: new unique-machine-id playbook
This playbook verifies that the contents of `/etc/machine-id` are
unique across multiple hosts.  It requires exactly two hosts that have
the same distribution and distribution version.

With this change, an attempt was made to start following the Openshift
Ansible best practices, as mentioned in issue #1.  This includes
checking that variables are defined before playbook/role/task
execution and a shift away from using `include:` statements.

The new approach is to use roles and symlink a top-level directory
into each individual test directory.  This has been found to work
quite well.