Improvements to the sanity test
The existing sanity test(s) have always felt lacking in terms of
functional tests that are executed.  This change attempts to improve
(or perhaps, replace) the existing sanity test(s).

This is the first cut at these types of tests.  The intent is to
expand the coverage of this set of tests to include additional tests
that are small and focused.  More complex functional tests can be
covered in separate playbooks which can be run after this.

Notable changes:
  - roles everywhere
  - hard commitment to Ansible version 2+
  - selectable Vagrant boxes for testing playbooks
  - verify that /tmp is functional after each reboot
  - verify that added users persist after upgrade
  - verify changes to /etc persist after upgrade
  - verify that directories/files in /var persist after
    upgrade/rollback
  - verify that users added after upgrade do not persist after
    rollback