roles: use ansible_hostname instead of ansible_fqdn
While using the `improved_sanity_test` playboook in our OpenStack
environment (when driven by Jenkins), we were seeing failures when trying to
verify changes made to `/etc/hosts`.  (We did not see the same
failures when running the playbook by hand.)

The root cause was the value of `ansible_fqdn` was changing after the
host was rebooted.  I don't fully understand why this was changing
after a reboot, but in order to work around this, we can just use
`ansible_hostname` in its stead.