Reset refspec in origin for alternate remotes
The subscription manager changes the remote in the refspec in the
origin file.  If the test system is pointing to other update streams
like autobrew builds, the refspec is overwritten and causes issues
with upgrading.  To fix this, the refspec is now saved when it is not
the default value and re-written after the subscription process.

This fix is particularly important for tests that deal with building
RHEL containers and upgrading to alternate remotes.  Building RHEL
containers requires entitlements from the host.  Upgrades to alernate
remotes requires the proper refspec in the origin file.