Add package layering test
Because roles and tasks cannot be be combined in Ansible, roles were created
for positive scenarios in package install and uninstall and its verifications
only.

Negative tests and tests with unexpected output are done afterwards in
the tasks portion of the playbook. Unfortunately there is a lot of repeated
code to validate output in the task section because we cannot combine roles.