Make each test a separate play
Reorganizing the playbook into separate plays to facilitate viewing
and triaging of tests.  Each playbook was given a tag so a user can
choose to run specific test(s).

There is still an issue with Ansible ignoring roles that are called
multiple times so some reboots were tagged with different names.