atomic_scan_verify: use upstream scanner
We were seeing failures on certain streams when trying to install the
`openscap` scanner from `registry.access.redhat.com`.  The version of
`skopeo` being used in these cases did not support v1 schema manifests
and that is the only version available on that registry.

Using the scanner from upstream alleviates this problem and gives us
further coverage in this space.

Oddly, only RHEL-based images are supported by the `openscap` scanner,
so the behavior has been changed to scan the RHEL7 image regardless if
the scanner used is upstream or downstream.

Finally, `atomic scan` no longer requires a TTY, so we can change
how that command is invoked.