docker_build_httpd:  add makecache.sh script
We have seen failures in the tests that appear to be related to
network issues (either in our OpenStack infra or in the Fedora infra)
when trying to install `httpd` while building the Docker image.

To combat this, we just steal the [env_make_cache()](https://github.com/jlebon/redhat-ci/blob/master/testrunner#L267)
routine from @jlebon and stick it in a dumb script.  Then we `ADD` to
the Docker image and `RUN` it before trying to install any packages.

This does add some extra instructions and time while building the
Docker image, but should help us avoid these network related failures.