Add pkg layering and admin unlock to sanity test
Adding the package layering and ostree admin unlock tests to the
improved sanity tests.  The roles that were created for package
layering and admin unlock were leverage for basic sanity tests
through upgrades and rollbacks.

There was an issue with package layering a package from the EPEL
repo and the etc_modify role where the EPEL repo was added into
/etc.  This caused upgrades to fail because it could not find the
GPG KEY for the EPEL repo.  The EPEL repo GPG key is now copied
in as a part of the etc_modify role to work around this issue.

One more additional change was adding retries to the rpm-ostree
upgrade because flaky responses from Fedora repo caused the test
to intermittently fail with Socket I/O Timeout.