Add docker(-latest) tests
This playbook adds some basic functionality tests for docker and
docker-latest.  It leverages much of the work that was done for
docker in the improved sanity tests.  There was one small change
to docker_run_httpd where the httpd container is started and
stopped again to provide coverage for the docker start command.

By default, the tests will configure the atomic host to run
docker-latest then run the suite of test.  There is a boolean
in vars.yml named g_docker_latest that can be set to false if
the tester wants to run the tests against the default docker
instance.