reboot: use ansible_host
The wait_for module requires a resolvable host. The inventory_hostname
variable however is not guaranteed to be resolvable. It may be an alias.
Use ansible_host instead, which is the actual resolvable hostname that
ansible uses to SSH, which is exactly what we want here.

This also makes sure that the reboot knowledge is only in
common/ans_reboot.yaml, rather than copied a bit everywhere. Add a
parameter for the case where the reboot sequence was already initiated.

Resolves: #82