repo: add .redhat-ci.yml
Now that we have a known working tree with the CAHC smoketested branch,
we can spin this around and use it to automatically test new additions
to the improved sanity tests.

Of course, if we're adding new tests that catch bugs that are already
present in the latest smoketested branch, the PR tester will fail. Thus,
this is not meant to be used with a merge bot, but rather a general
indication of how sane a PR is.