Add runc tests
This check-in add a playbook to test runc functionality.  The playbook
sets up the environment needed to run an upstream runc python test on
Atomic Hosts.

The origin plan was to write the runc python test in Ansible so we
would have a consistent testing strategy.  However, there was no straight
forward way to manipulate JSON files in Ansible to test the various
runc options.  Ideally the test would not need a modified ostree
(package layering and admin unlock).

I did run into a problem with running the upstream test through Ansible.
The test requires a tty so I used the python command line workaround.
The problem is that the workaround would always return 0 even if the
script failed so I had to check for errors in stdout.  There are checks
for the most common errors that occurred when running the test but it may
not be all inclusive.