add a test for rhbz 1421867
Our internal testing found that trying to use `rpm-ostree status` on a
RHELAH host that was not registered via `subscription-manager` would
fail. See [RHBZ
1421867](https://bugzilla.redhat.com/show_bug.cgi?id=1421867)

To test for this, I've re-arranged when we use `rpm-ostree status` and
added a role that removes any subscriptions from a host and drops an
`unconfigured-state` field into the origin file (optionally).