expand SELinux tests
This commit expands the scope of our SELinux checks to include the
validation of the SELinux labels on multiple key processes and files.

It also checks that any files on the filesystem do not have the
default labels `default_t` or `file_t`.

Finally, it includes a simple test to validate that changes made
to an SELinux boolean will persist through the upgrade/reboot cycle.
This particular test covers [RHBZ#1381588](https://bugzilla.redhat.com/show_bug.cgi?id=1381588)