tests: remove unmaintained tests
Let's face reality and remove these tests that have not been
maintained or even ran since this repo was created.  Some of the tests
are covered by the `improved-sanity-tests` and others are run manually
as part of regular releases.

And since this is `git` they are always just a revert away, in case we
decide we have made a mistake.