Handle None case for ignore_errors
ignore_errors is set to None when it is passed into v2_runner_on_failed
which will not capture any failures except when the calling function
does not pass an ignore_errors parameter.  I swapped it to capture the
failure in all cases except when ignore_errors is explicitly set to
true.