semanage_add_verify: fix to run on all platforms
Previously we were trying to change the `fcontext` object type on a
specific directory based on if there was a `kubelet` package
installed.  This was incorrect, as I don't think any platform/stream
had that package installed.

This re-tools the role to modify an known object type to a new value
and compares the result at the end.  No uncessary checks and should
work on all streams.

Fixes #99