new test for validating base images with Atomic Host
(That's a bad summary of the changes, but I couldn't do better in less
than 70 characters.)

This test intends to validate that any changes to the Atomic Host
platform will not break the usage of any 'reasonably popular' base
images.  This is accomplished by building an `httpd` container on top
of each base image, running the resulting image, and verifying the
content returned from the container is correct.

I chose 'popular' base images from Docker Hub based on the number of
pulls.  I also threw in the Fedora images from the new Fedora
registry and the RHEL images from the Red Hat registry.