Change to more complex rpm-ostree install package
A new feature was added to allow non-root packages to be layered
using rpm-ostree install.  Instead of using a simple package like
wget, the httpd package  tests a broader code path because it has
multiple dependencies, creates an apache user, and is a non root
package.

The default settings for httpd conflicted with the docker container
tests in improved sanity tests so it was changed to port 8080 in the
httpd_start role.  The httpd service is validated by retrieving the
default apache page.