Consistently use .yml as the suffix for YAML files
The repo had files with both `.yaml` and `.yml` suffixes, which could
lead confusion when trying to include files as part of a playbook.
This is part of the OpenShift Ansible style guide, which we try to
stick to as much as possible.

As part of the rename, all references or `include:` statements were
updated to point to the correct files.

There may have been a few typos corrected or extra white spaces
removed along the way.

This should be merged after #117.