Fix metadata cache generation for RHEL case
We have seen instances on RHEL hosts where the `makecache.sh` script
inside the `docker build` process tries to generate the cache for every
repo a subscription provides.  Not only is this costly in terms of
time required to generate the cache, but it is also incorrect since we
explicitly disable all the repos and re-enable only those we care about
during the subscription phase.

To work around this, a copy of the `makecache.sh` script was made with
RHEL specific instructions to use only certain repos.