Add status table to repo landing page
This commit adds a status table of the latest improved sanity test
runs for the streams we test.  It uses shields.io badges containing
the version and result of the test run.  The badges are not directly
from shields.io, instead badges and logs are saved in a s3 bucket.