Fix k8-cluster test for Fedora
The k8-cluster test was not run against Fedora for quite some time.
The test was failing and after some debugging it turned out that the
python package was removed from the Fedora latest container causing
cgi-bin/action in the web server to not be able to connect with the
mariadb instance.  Python3 is still available in the Fedora container.

There were two solutions: convert the action script to python3 or
install python in the Dockerfile.  I chose the latter because the action
script was used in documentation and wanted to keep it consistent.

The Dockerfile and content was compressed in a tgz file which makes it
difficult to make changes to the content.  The tgz was extracted to
facilitate future changes to the content.  The appropriate changes were
made to the role to support the new file structure.