roles: modify 'selinux_verify' to be distro aware
This changes how the `selinux_verify` role will check the SELinux file
and process labels.  The main improvement is the ability to create
distro-specific checks of the labels.

Instead of one master file that contains all the files and processes
we want to check, there is now two locations where this information
is defined.  There is a 'common' file that contains the files/processes
that can be found on all supported streams.  And there are distro-specific
(per major-release) files that define files/processes that which don't
land in the 'common' set.  This brings in great flexibility for
supporting new versions of streams, as well as the ability to expand
the files/processes we want to check.

This change also does away with the complexity of the
`with_subelements` method that was used before and uses `with_items`
which is easier to understand.  The drawback to this approach is some
duplication of key/values, but I think the overall gain is worth it.

Closes #133