Cleanup after test completion
If we want to run the tests in succession, we should cleanup what the
tests have left around.  This typically includes docker
images/containers, kubernetes artifacts, and layered packages.

This change attempts to cleanup after most* tests and return the host
under test to its original state.  This includes two new roles that
assist in the cleanup of the host.

* The improved-sanity-test and system-containers test were not
addressed in this change, but should be addressed in the future.