Use empty string for extra args if undefined
In #153 support was added for passing extra arguments to the
subscription manager command.  If no value was defined in the
subscription csv for 'extras' the lookup command would return []
which would then be added to the end of the subscription manager
command and fail.  Since the return from the lookup is [], the
jinja2 default filter of empty string did not get applied.

The fix I came up with was to set extra_args to the extras lookup
only if the extra lookup is not [].  Otherwise it will set it to
empty string.  This preserves backwards compatability to the old
subscription csv format for people that are already using the test.