Simplify rpm install verification in improved sanity test
Addresses issue #175.  The previous method of verifying rpm installs
relied scraping HTML to determine the filename for a RPM before it
could install it.  There was also a single point of failure pointing
to dl.fedoraproject.org instead of the mirror site.

This PR simplifies the verification in the improved sanity test by
using the epel-release package in the mirror site which always has
the same URL.