Generalize rpm_install role
Instead of having a role specifically to install packages from EPEL,
it was suggested to have a generic role that installs any package if
given a url.  In order to preserve the functionality of the origin role,
a separate role (epel_rpm_url) was created that returns the full url of
a package after scraping HTML.  These two roles replace the current
unctionality of rpm_install.

After the role changes, the improved-sanity-test and admin-unlock tests
needed to be modified to work with the new roles.