Use correct git diff to determine tests to run
The two dot notation used originally does not work the same way in
git diff as other git commands.  Git diff expects two endpoints as
opposed to a range.  This was causing the diff to return files that
were not a part of the PR to be included in the automatic tests that
are run when there is a new PR.