i-s-t: cleanup any extra deployments at the end
When running the `improved-sanity-test` using `--tags cloud_image` the
system is left with a 'hotfix' deployment in the waiting.  If we want
to use the system again to run additional tests, the nice thing to do
is to cleanup after ourselves.

In the case of normal execution, where the system starts at HEAD-1,
upgrades, than rolls back to HEAD-1, we will end up cleaning the
deployment at HEAD.  I'm of the opinion this isn't a deal-breaker.