roles: implement retries and remove version check
We no longer need to check the version of the `atomic` command in
order to determine how to list images with the command, since all the
platforms we test against are at version 1.12+ of `atomic`.

Additionally, I added a retry loop on the `atomic pull` in order to
mitigate any failures due to network problems.

This address a small sliver of #156