add support for removing all installed packages
Sometimes you just want all the installed packages removed, so this
introduces support for the `all` keyword in the `rpm_ostree_uninstall`
role.  It is not 100% bulletproof, since it is possible a user could
use the `all` keyword when there is nothing installed and then the
role will probably blow up.  But I'm hoping that we are better than that.

There could be additional work done in the
`rpm_ostree_uninstall_verify` role to support the `all` keyword, too.
But I was unable to arrive a solution that would both verify the
packages were uninstalled and the binaries were removed.