Add etcd system container test to I-S-T
etcd_sys_container_install role:
This check-in is a simple system container test for the improved sanity test (IST)
Because the IST is comprised of only roles (and cannot be mixed with tasks), the
etcd_sys_container_install role was made specifically to validate the install
and configuration of etcd for IST.

improved sanity test:
In the IST, the etcd system container is installed and configured pre-upgrade
and checked post upgrade for its existence.  Then it is removed and re-installed
to test the system container installation on HEAD.  On a cloud_image test, the
installation, configuration, and uninstallation is only checked once in the post
upgrade section. This was done to ensure we have some coverage on the upgrade
scenario and when running the test on a cloud_image with no upgrade.

command / shell role:
I added these two simple roles because sometimes you just want to run a command
and check the output when you are in a role block.  The command role was used
to quickly validate the etcd service is running.