i-s-t: use sythensized upgrades to test HEAD
Because the current improved sanity test requires the test system
to be at HEAD-1, it made it more complicated to figure out when
issues were introduced.  Consumers of the test results would have
to wait until HEAD+1 to see if patch worked.  This commit aims
to resolve these issues by using a synthesized upgrade provided
by @jlebon.

The synthesized upgrade creates a local branch of the current ostree
then creates a commit with no changes.  At this point running an
upgrade would upgrade to an ostree that does not include
packages that are added/upgraded, but the trade-off is worth it for
better CI tests and less user confusion.  Users can see when their
changes causes breakage and when their patches fix them.

Another issue with using this approach is that the synthesized
upgrades is that the system must be upgraded to the local branch
right away.  Creating the local repo later in the test does not
preserve admin/unlock and layered packages that are added during
the test.

By default, the improved sanity test will run with synthesized
upgrades.  A tag named 'synthesized_upgrade' was added to the
extra roles so users can run without simulated upgrades.  Just
run with option --skip-tags synthesized_upgrade.