docker_build_httpd: add retry logic
As part of the effort to enable retries on network operations, this
turns on retries when doing a `docker build`.  This is because the
operation has to fetch metatdata and RPM data from the network as part
of the build, so might as well try a few times in case the network
sucks.