Add support for checking the journal for strings
This introduces two roles that can be used for checking the journal of
a host for particular strings.

`journal_search`:  A role that will search the journal for a specific
                   string supplied to the role.  Optionally can specify
                   that the role fails when the string is found.

`journal_fatal_msgs`:  A 'meta' role that calls out to `journal_search`
                       to look for certain phrases that are considered
                       fatal to the overall test.

The `journal_fatal_msgs` role is used in the `improved-sanity-test` to
inspect the state of the journal before and after the host is
rebooted during the test duration.

Closes #186