roles: add ability to check RPM signatures
The three platforms which we support with these tests have periodic
official releases where the RPMs are all signed by a well known GPG
key.  When testing these releases/streams, we want to be able to
automtically verify that the RPMs have been signed with the correct
GPG key.

This role adds support for verifying that the RPMs are correctly
signed.  In its current implementation, we only support checking for a
single GPG signature, which is how the RPM content has been
historically signed.  In other words, none of the platforms currently
sign their RPMs with multiple keys for an official release.

This role has also been added to the `improved-sanity-test`.