roles:  fix b_d_s_f to work with local branches
If the `booted_deployment_set_fact` role is used after the host has
been upgraded/rebased to a local branch, it will currently fail
because the refspec has no `remote` portion.

This fixes the role to work with a local branch by checking for the
`remote:refspec` syntax (just the presence of the `:` actually) and
then setting up the facts appropriately.