Save original deployment info for later use
It's possible that through the course of testing that the original
deployment information is lost (more acccurately, not easily
available) via upgrades/rebases.  In order to return the host under
test to its original state, we need a way to retain the original
deployment information to be recalled later.

This problem is solved by writing out the deployment information
(`checksum`, `refspec`, and `remote`) to an INI-style file in the
local playbook directory.  This file can then be read in by another
role that will setup variables to be used by the `rpm_ostree_rebase`
role.

The solution is admittedly a bit specific to this problem/variables,
but I think could be modified in the future to be a bit more
versatile.