roles: refactor rpm_ostree_rebase to support more variations
The `rpm-ostree rebase` command supports multiple variations and this
change aims to support them with the `rpm_ostree_rebase` role.

Basic variations that are suported:
  - remote:branch
  - remote:branch commit
  - :local-branch
  - :local-branch commit

The role employs a bit of input checking to make sure there is
something usable, but it is still possible for a careless user to
provide inputs that will make it blow up.