Archive original deployment info to YAML file
This commit is a re-work of my the first approach to saving off the
original deployment info.  Here the information is templated out to a
YAML file and then loaded dynamically later on.

A generic role called `load_variables` has been provide that can be
used to load in variables from a YAML file.

This also removes the too-specific `original_refspec_set_fact` role.